Add a default Chinese mode to all tables
ibus-table uses the following priority to select the Chinese mode:

1) User setting (can be set in the setup tool or the property menu)
2) The value of LANGUAGE_FILTER in the table source
3) The locale

Getting the default for the Chinese mode from the locale is only a
last ditch guess. A user who wants to use a table for traditional
Chinese does not necessarily run his system in a traditional Chinese
locale.  It is better if each table already specifies a default
suitable for that table.